requirements.txt
google-genai
python-dotenv
chromadb
google-api-core